aiservice/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
opencv-python
numpy
pillow
pytesseract
